[REF] Refactor to go back to a single component approach
