--- packages/ng/demo/app/selects/expert-select/expert-select.component.html
@@ -1,18 +1,0 @@
-<h1>With Options</h1>
-<div class="textfield mod-select mod-material" [class.mod-remove]="_canRemove">
-	<input [luSelect]="selectRefOptions"
-		[(ngModel)]="itemOptions"
-		class="textfield-input"
-		name="itemOptions"
-		luEmpty
-		clearable="true"
-		(canremove)="canRemove($event)">
-	<label class="textfield-label" for="itemOptions">my label</label>
-	<lu-select-picker #selectRefOptions>
-		<lu-select-option *ngFor="let option of options" [value]="option">{{option.name}}</lu-select-option>
-	</lu-select-picker>
-	<button class="actionIcon" (click)="clear()">
-		<i class="lucca-icon">cross</i>
-	</button>
-</div>
-<code class="code mod-block">{{ itemOptions | json }}</code>
