--- packages/ng/demo/app/selects/expert-select/expert-select.component.scss
@@ -1,52 +1,0 @@
-@import "node_modules/@lucca-front/scss/src/theming.overridable";
-@import "node_modules/@lucca-front/icons/src/mixins";
-
-.textfield {
-	.actionIcon {
-		position: absolute;
-		font-size: .75rem;
-		bottom: .45rem;
-		right: 2.75 * _component("textfield.input.padding-horizontal");
-		padding: 0;
-		width: 1rem;
-		height: 1rem;
-		display: none;
-	}
-}
-
-.textfield-input {
-  overflow: hidden;
-  overflow: hidden;
-  text-overflow: ellipsis;
-	white-space: nowrap;
-}
-
-
-// MODS
-// ▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒
-
-.textfield {
-
-
-	&.mod-remove {
-		.textfield-input {
-			padding-right: 4.5 * _component("textfield.input.padding-horizontal");
-		}
-
-		.actionIcon {
-			display: block;
-		}
-	}
-
-	&.mod-material {
-		.actionIcon {
-			right: 2 * _component("textfield.input.padding-horizontal");
-		}
-
-		&.mod-remove {
-			.textfield-input {
-				padding-right: 4 * _component("textfield.input.padding-horizontal");
-			}
-		}
-	}
-}
