+++ packages/ng/demo/app/selects/selects.component.html
@@ -10,10 +10,3 @@
 <demo-example-box [snippets]="snippets" demo="simple">
 	<demo-simple-select></demo-simple-select>
 </demo-example-box>
-
-<h3>Expert</h3>
-<demo-example-box [snippets]="snippets" demo="expert">
-	<demo-expert-select></demo-expert-select>
-</demo-example-box>
-<!--
--->
