+++ packages/ng/demo/app/selects/simple-select/simple-select.component.html
@@ -1,11 +1,16 @@
 <h1>Component With Options</h1>
 
-<lu-select [(ngModel)]="itemSelect"
+<div class="textfield mod-material">
+
+	<lu-select [(ngModel)]="itemSelect"
 	class="text-field-input"
 	mod="mod-material"
 	placeholder="my label"
+	name="itemOptions"
 	clearable="true">
 	<lu-select-option *ngFor="let option of options" [value]="option">{{option.name}}</lu-select-option>
 </lu-select>
+<!--<label class="textfield-label" for="itemOptions">my label</label>-->
+</div>
 
 <code class="code mod-block">{{ itemSelect | json }}</code>
