+++ packages/ng/src/app/select/directive/select.directive.spec.ts
@@ -28,8 +28,7 @@ const globalOptions = [{
 }];
 
 @Component({
-	template: `<input [luSelect]="pickerRef"
-			[(ngModel)]="item"
+	template: `<div [luSelect]="pickerRef"
 	>
 	<lu-select-picker #pickerRef>
 		<lu-select-option *ngFor="let option of options" [value]="option">{{option.name}}</lu-select-option>
@@ -37,7 +36,6 @@ const globalOptions = [{
 })
 export class WrapperSelectDirective {
 	options = globalOptions;
-	item = globalOptions[0];
 }
 
 describe('LuSelectDirective', () => {
@@ -65,7 +63,7 @@ describe('LuSelectDirective', () => {
 				this.fixture = TestBed.createComponent(WrapperSelectDirective);
 				this.fixture.detectChanges();
 				this.wrapper  = this.fixture.debugElement.query(By.directive(LuSelectDirective));
-				this.luSelect = this.wrapper.injector.get(LuSelectDirective) as LuSelectDirective<any>;
+				this.luSelect = this.wrapper.injector.get(LuSelectDirective) as LuSelectDirective;
 		});
 
 		it('It should reference the correct picker', () => {
@@ -82,91 +80,6 @@ describe('LuSelectDirective', () => {
 			expect(this.luSelect.popover.luOptions$.value.length).toBe(globalOptions.length);
 
 
-		});
-
-		it('It should reference has the right value', () => {
-
-			// Arrange
-			this.fixture.whenStable().then(() => {
-				// Act
-				this.fixture.detectChanges();
-
-				// Assert
-				expect(this.luSelect.value).toBe(globalOptions[0]);
-			});
-
-		});
-
-		it('It should display the right text', () => {
-
-			// Arrange
-			const divDirective = this.fixture.debugElement.query(By.css('div'));
-
-			this.fixture.whenStable().then(() => {
-				// Act
-				this.fixture.detectChanges();
-				const viewValue = this.luSelect.popover.luOptions$.value[0].viewValue;
-
-				// Assert
-				expect(divDirective.nativeElement.textContent).toBe(viewValue);
-			});
-
-		});
-
-		it('It should not be clearable by default', () => {
-			// Arrange
-			this.fixture.whenStable().then(() => {
-				// Act
-				this.fixture.detectChanges();
-
-				// Assert
-				expect(this.luSelect.clearable).toBe(false);
-			});
-		});
-
-		it('It should allow empty value if clearable', () => {
-				// Arrange
-
-				this.fixture.whenStable().then(() => {
-					// Act
-					this.fixture.detectChanges();
-					this.luSelect.clearable = true;
-					this.luSelect.value = null;
-					this.fixture.detectChanges();
-
-					// Assert
-					expect(this.luSelect.value).toBeUndefined();
-				});
-		});
-
-		it('It should not allow empty value if not clearable', () => {
-			// Arrange
-
-			this.fixture.whenStable().then(() => {
-				// Act
-				this.fixture.detectChanges();
-				this.luSelect.value = null;
-				this.fixture.detectChanges();
-
-				// Assert
-				expect(this.luSelect.value).toBe(globalOptions[0]);
-			});
-		});
-
-		it('It should emit an event when we can clear the value', () => {
-			// Arrange
-			spyOn(this.luSelect.canremove, 'emit');
-
-			this.fixture.whenStable().then(() => {
-				// Act
-				this.fixture.detectChanges();
-				this.luSelect.clearable = true;
-				this.luSelect.value = null;
-				this.fixture.detectChanges();
-
-				// Assert
-				expect(this.luSelect.canremove.emit).toHaveBeenCalled();
-			});
 		});
 
 });
