+++ packages/ng/src/app/select/select.component.ts
@@ -4,11 +4,15 @@ import {
 	Input,
 	EventEmitter,
 	forwardRef,
+	Renderer2,
 	ElementRef,
+	HostListener,
+	HostBinding,
 	OnInit,
 	OnDestroy,
 	QueryList,
 	ContentChildren,
+	ViewContainerRef,
 	ViewEncapsulation,
 	ViewChild,
 } from '@angular/core';
@@ -31,6 +35,24 @@ import {startWith} from 'rxjs/operators/startWith';
 import {takeUntil} from 'rxjs/operators/takeUntil';
 import {LuSelectOption} from './option/select-option.component';
 
+/** KeyCode for End Key */
+const END = 'End';
+/** KeyCode for Home Key */
+const HOME = 'Home';
+/** KeyCode for Escape Key */
+const ESCAPE = 'Escape';
+/** KeyCode for Backspace Key */
+const BACKSPACE = 'Backspace';
+/** KeyCode for Delete Key */
+const DELETE = 'Delete';
+/** KeyCode for Enter Key */
+const ENTER_KEY = 'Enter';
+/** KeyCode for ArrowUp Key */
+const UP_KEY = 'ArrowUp';
+/** KeyCode for ArrowDown Key */
+const DOWN_KEY = 'ArrowDown';
+const TAB = 'Tab';
+
 /**
  * The component that provides available options from the api with the currently inputed text
  */
@@ -44,24 +66,54 @@ import {LuSelectOption} from './option/select-option.component';
 	],
 	styleUrls: ['./select.component.scss'],
 })
-export class LuSelect<T> implements ControlValueAccessor, AfterContentInit, OnInit, OnDestroy {
+export class LuSelect<T>
+implements ControlValueAccessor, AfterContentInit, OnInit, OnDestroy {
 
 	/** Emits whenever the component is destroyed. */
 	private _destroy$ = new Subject<void>();
 	/** inner validator */
 	protected _validator: ValidatorFn | null;
+
 	/** True if the the component allow the clear of data  */
 	protected _canRemove = false;
 	/** The value of the select */
 	get value(): T | null {
-		return this._field ? this._field.value : null;
+		return this._value;
 	}
+	/** Set the value, an event (canremove) will be sent if the directive is clearable */
 	set value(value:  T | null) {
-		this.writeValue(value);
+		let valueTemp = value;
+		if (valueTemp === null
+				&& !this.clearable
+				&& this._picker
+				&& this.luOptions
+				&& this.luOptions.length > 0) {
+			valueTemp = this.luOptions.first.value;
+		}
+		const lastValue = this._value;
+		this._value = valueTemp;
+		// emit change
+		if (!this._same(lastValue, valueTemp)) {
+			this._cvaOnChange(valueTemp);
+			if (this.clearable) {
+				this._emitClearable();
+			}
+			// Transfer the information to popover
+			this._picker.selectOption(value);
+			this._picker.find(value).subscribe(selectOption => {
+				this._selectOption = selectOption;
+				// render
+				this.render(selectOption);
+			});
+		}
 	}
 
+	// Inner values
+	protected _value: T | null;
+	protected _selectOption: LuSelectOption<T> | null;
+
 	// Inner Children
-	@ViewChild(LuSelectDirective)_field: LuSelectDirective<T>;
+	@ViewChild(LuSelectDirective) _field: LuSelectDirective;
 	@ViewChild(LuSelectPicker) _picker: LuSelectPicker<T>;
 	/**
 	 * List of LuSelectOptions
@@ -84,15 +136,24 @@ export class LuSelect<T> implements ControlValueAccessor, AfterContentInit, OnIn
 	private _itemValidator: ValidatorFn = (): ValidationErrors | null => {
 		return null;
 	}
+	private _cvaOnChange: (value: T) => void = () => {};
 	_onTouched = () => {};
 
-	constructor(private element: ElementRef
+	constructor(
+		protected _elementRef: ElementRef,
+		protected _renderer: Renderer2,
 	) {
 	}
 
 	// Life Cycle methods
 	ngOnInit() {
 		this._validator = Validators.compose([this._itemValidator]);
+		this._renderer.setAttribute(this._elementRef.nativeElement, 'tabindex', '0');
+		this._picker.itemSelected
+		.subscribe(item => {
+			this.value = item ? item.value : undefined;
+			this._field.closePopover();
+		});
 	}
 	ngOnDestroy() {
 		this._destroy$.next();
@@ -106,6 +167,7 @@ export class LuSelect<T> implements ControlValueAccessor, AfterContentInit, OnIn
 				this._picker.luOptions$.next(this.luOptions.toArray());
 			}
 		});
+
 		Promise.resolve().then(() => {
 			if (!this.clearable && !this.value) {
 				if (this.luOptions.length === 0) {
@@ -118,13 +180,11 @@ export class LuSelect<T> implements ControlValueAccessor, AfterContentInit, OnIn
 
 	// From ControlValueAccessor interface
 	writeValue(value: T) {
-		if (this._field) {
-			this._field.value = value;
-		}
+		this.value = value;
 	}
 	// From ControlValueAccessor interface
 	registerOnChange(fn: any) {
-		this._field.registerOnChange(fn);
+		this._cvaOnChange = fn;
 	}
 	// From ControlValueAccessor interface
 	registerOnTouched(fn: any) {
@@ -149,6 +209,88 @@ export class LuSelect<T> implements ControlValueAccessor, AfterContentInit, OnIn
 		this.value = option ? option.value : null;
 	}
 
+	// render/display
+	protected render(value = this._selectOption) {
+		this._elementRef.nativeElement.value = this.display(value);
+	}
+	protected display(value: LuSelectOption<T> | null): string {
+		if (!value) {
+			return '';
+		}
+		return value.viewValue;
+	}
+
+	// host listening
+	@HostListener('keydown', ['$event'])
+	onKeydown($event) {
+		switch ($event.key) {
+			case ESCAPE:
+				if (this._field.popoverOpen) {
+					this._field.closePopover();
+				}
+				break;
+			case DELETE:
+			case BACKSPACE:
+				if (this.clearable) {
+					this.value = null;
+				}
+				break;
+			case HOME:
+			$event.preventDefault();
+			return this._picker.onHomeKeydown(this._field.popoverOpen);
+			case END:
+			$event.preventDefault();
+			return this._picker.onEndKeydown(this._field.popoverOpen);
+			case ENTER_KEY:
+			return this._field.popoverOpen ? this._picker.onEnterKeydown() : this._field.openPopover();
+			case DOWN_KEY:
+			$event.preventDefault();
+			return this._picker.onDownKeydown(this._field.popoverOpen);
+			case UP_KEY:
+			$event.preventDefault();
+			return this._picker.onUpKeydown(this._field.popoverOpen);
+			case TAB:
+			break;
+			default:
+			$event.preventDefault();
+		}
+	}
+
+	@HostListener('blur', ['$event'])
+	blur(e) {
+		this._onTouched();
+		this._field.closePopover();
+	}
+
+	@HostListener('click')
+	clicked() {
+		this._field.togglePopover();
+		if (this._field.popoverOpen) {
+			this._picker.search(this._strValue);
+		}
+	}
+
 	// Utilities
+
+	protected get _strValue(): string {
+		return this._elementRef.nativeElement.value as string;
+	}
+	private _emitClearable() {
+		this.canRemove(this.clearable && !!this.value);
+	}
+
+	/** detect via JSON.stringify if both value are equals */
+	protected _same(oldItem: T, newItem: T) {
+		if (oldItem === newItem) {
+			return true;
+		}
+		if (!oldItem && !newItem) {
+			return true;
+		}
+		if (!oldItem || !newItem) {
+			return false;
+		}
+		return JSON.stringify(oldItem) === JSON.stringify(newItem);
+	}
 
 }
